Refactor onboarding management by introducing a dedicated onboarding storage module for desktop and extension clients. Migrate onboarding-related functions from the core storage interface, enhancing code organization and simplifying the storage structure. Remove obsolete onboarding functions from the storage index files.
